+++ toucan_connectors/mysql/mysql_connector.py
@@ -177,6 +177,12 @@ class MySQLConnector(
         title="Charset",
         description='Character encoding. You should generally let the default "utf8mb4" here.',
     )
+    charset_collation: str | None = Field(
+        None,
+        title="Charset collation",
+        description="The charset's collation for the connections to the server."
+        "Only set it here if your tables do not use your server's default value.",
+    )
     connect_timeout: int | None = Field(
         None,
         title="Connection timeout",
@@ -278,6 +284,7 @@ class MySQLConnector(
             "connect_timeout": self.connect_timeout,
             "conv": conv,
             "cursorclass": cursorclass,
+            "collation": self.charset_collation,
         }
         # remove None values
         return {k: v for k, v in con_params.items() if v is not None}